+++ pyqpp/include/pyqpp/classes/qengine_bind.hpp
@@ -93,10 +93,7 @@ void declare_QEngineT(py::module& m) {
              "True if post-selection was successful (or absent), false "
              "otherwise")
         .def("reset", &QEngineT<T>::reset, "Resets the engine",
-             py::arg("reset_stats") = true,
-             py::arg("ensure_post_selection") = true,
-             py::arg("max_post_selection_reps") =
-                 std::numeric_limits<idx>::max())
+             py::arg("reset_stats") = true)
         .def("reset_stats", &QEngineT<T>::reset_stats,
              "Resets the collected measurement statistics hash table")
         .def("set_dit", &QEngineT<T>::set_dit,